requirements.txt
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
colorama==0.4.4
coverage==5.5
decorator==5.0.9
future==0.18.2
iniconfig==1.1.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.1
jedi==0.18.0
kafka-python==2.0.2
lazy-object-proxy==1.6.0
mando==0.6.4
matplotlib-inline==0.1.2
mccabe==0.6.1
packaging==20.9
parso==0.8.2
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.19
py==1.10.0
pycodestyle==2.7.0
Pygments==2.9.0
pylint==2.8.3
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv==0.18.0
radon==5.0.1
six==1.16.0
snakeviz==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
